requirements.txt
numpy==1.25.2
opencv-python==4.8.0.74
Pillow==10.0.0
pydub==0.25.1
PyQt6==6.5.2
PyQt6-Qt6==6.5.2
PyQt6-sip==13.5.2
